Add recent project picker
