+++ crates/zed/src/main.rs
@@ -121,6 +121,7 @@ fn main() {
         vim::init(cx);
         terminal::init(cx);
         theme_testbench::init(cx);
+        recent_projects::init(cx);
 
         cx.spawn(|cx| watch_themes(fs.clone(), themes.clone(), cx))
             .detach();
